backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.9",
    "django-cors-headers>=4.9.0",
    "django-otp>=1.6.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "joblib>=1.5.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "redis>=7.1.0",
    "reportlab>=4.4.5",
    "scikit-learn>=1.7.2",
    "whitenoise>=6.11.0",
]
